Use LoginOutputData to follow CA architecture.
