+++ src/main/java/com/studyarc/use_case/login/LoginInputBoundary.java
@@ -1,5 +1,5 @@
 package com.studyarc.use_case.login;
 
 public interface LoginInputBoundary {
-    LoginResult login(LoginInputData loginInputData);
+    LoginOutputData login(LoginInputData loginInputData);
 }
